cookbook/requirements.txt
Babel==2.9.0
certifi==2020.12.5
chardet==4.0.0
Delorean==1.0.0
humanize==3.3.0
idna==2.10
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
tzlocal==2.1
urllib3==1.26.4
parse==1.8.2
PyYAML==3.12
beautifulsoup4==4.6.0
feedparser==5.2.1 
